refactor(KERN-1560): split extract-filters into focused submodules
extract-filters/ folder:
- extract-where.ts (280 lines): WHERE extraction, comparison, BETWEEN,
  IN/NOT IN, LIKE/ILIKE, OR/AND logic
- extract-having.ts (95 lines): HAVING extraction (all-or-nothing)
- filter-schema.ts (90 lines): ensureFilterColumnInSchema, type inference
- index.ts: barrel re-export

work-item: KERN-1560
